ARENA-18218 - Replace custom Authorization Header in the Ehub Provider Interface with JWT

+++ authinfo/src/main/java/com/axiell/authinfo/jwt/JwtAuthHeaderParser.java
@@ -92,7 +92,7 @@ public class JwtAuthHeaderParser implements IAuthHeaderParser {
         if (value instanceof String) {
             return builder.withClaim(key, String.class.cast(value));
         } else if (value instanceof Long) {
-            return builder.withClaim(key, (int) Long.class.cast(value).longValue());
+            return builder.withClaim(key, Long.class.cast(value).intValue());
         } else if (value instanceof Integer) {
             return builder.withClaim(key, Integer.class.cast(value));
         } else if (value instanceof Boolean) {
